requirements.txt
django
djangorestframework
djangorestframework-simplejwt
markdown
django-filter
